// src/lib/lemmy-js-client/src/types/ListUserChatRoomsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import {ChatRoomResponse} from "./ChatRoomResponse";
import type { PaginationCursor } from "./PaginationCursor";

/**
 * The response for listing the user's chat rooms.
 */
export type ListUserChatRoomsResponse = {
    rooms: ChatRoomResponse[];
    /** the pagination cursor to use to fetch the next page */
    nextPage?: PaginationCursor;
    /** the pagination cursor to use to fetch the previous page */
    prevPage?: PaginationCursor;
};